Background noise variance from observed data is now added to variance terms of both instrumental distribution and forward mapping (added as a scalar after all other variance adjustments)
When using the forward map, we improve the instrumental distribution covariance matrix by taking the point-estimate of the best release funtion, adjusting it slightly and mapping it forward through the model, to estimate a covariance matrix on data space, DV.  DV is then adjusted slightly.  We then  solve DV = A MV A^t for the model-space covariance matrix MV, where A is the forward map.  The instrumental distribution using this MV should then more closely resemble the true likelihood function of the forward map.

The default adjustments to the release function are currently to put a positive floor on it so as to create variance in otherwise silent parts of the matrix.  The adjustements to DV are to scale up the diagonal terms of the variance matrix.  Both of these adjustments make for larger MV, and thus more liberal instrumental sampling.